https://github.com/Yakup-TBR/jurnal5_1302210113

+++ jurnal5_1302210113/Penjumlahan.cs
@@ -15,10 +15,12 @@ namespace jurnal5_1302210113
             dynamic b = B;
             dynamic c = C;
 
-            float total = a + b + c;
+          
+
+            double total = a + b + c;
 
             System.Console.WriteLine(total);
 
-        }
+        } 
     }
 }
